+++ plugin/kvm/src/main/java/org/zstack/kvm/KVMConstant.java
@@ -128,6 +128,9 @@ public interface KVMConstant {
     String KVM_VERIFY_ENVELOPE_KEY_PATH = "/host/key/envelope/checkEnvelopeKey";
     String KVM_ENSURE_SECRET_PATH = "/host/key/envelope/ensureSecret";
 
+    /** HTTP timeout in seconds for envelope key sync (verify/create/rotate/get) to agent. */
+    long ENVELOPE_KEY_HTTP_TIMEOUT_SEC = 5L;
+
     String KVM_HOST_FILE_DOWNLOAD_PATH = "/host/file/download";
     String KVM_HOST_FILE_UPLOAD_PATH = "/host/file/upload";
     String KVM_HOST_FILE_DOWNLOAD_PROGRESS_PATH = "/host/file/progress";
