+++ test/src/test/resources/springConfigXml/HostSecretEnvelopeCryptoTestExtension.xml
@@ -7,9 +7,9 @@
     http://zstack.org/schema/zstack
     http://zstack.org/schema/zstack/plugin.xsd">
 
-    <bean id="MockHostSecretEnvelopeCryptoExtensionPoint" class="org.zstack.test.integration.kvm.host.MockHostSecretEnvelopeCryptoExtensionPoint">
+    <bean id="HostSecretEnvelopeCryptoTestExtension" class="org.zstack.test.integration.kvm.host.secret.HostSecretEnvelopeCryptoTestExtension">
         <zstack:plugin>
-            <zstack:extension interface="org.zstack.header.secret.HostSecretEnvelopeCryptoExtensionPoint" />
+            <zstack:extension interface="org.zstack.kvm.HostSecretEnvelopeCryptoExtensionPoint" />
         </zstack:plugin>
     </bean>
 </beans>
